Add debug logging in demo.py to trace total_steps value
Added debug output in the visualization worker to see what value of
total_steps is actually being received and displayed. This will help
identify if the value is correct when sent but wrong when displayed,
or if it's already wrong when received.

Removed previous debug log in speedrun_experiment.py since you confirmed
total_steps looks correct there.